Move inactiveSymbols sync-schema to separate file
